Modulo pesquisa e melhorias de layout
