+++ e2e/src/produto.e2e-spec.ts
@@ -2,17 +2,17 @@ import { ProdutoPage } from "./pageobject/produto.po";
 import { ComponentInterface } from './pageobject/common.po';
 import { by, element, browser, protractor } from 'protractor';
 
-var origFn = browser.driver.controlFlow().execute;
-browser.driver.controlFlow().execute = function() {
-   var args = arguments;
+// var origFn = browser.driver.controlFlow().execute;
+// browser.driver.controlFlow().execute = function() {
+//    var args = arguments;
  
-   // queue 100ms wait
-   origFn.call(browser.driver.controlFlow(), function() {
-     return protractor.promise.delayed(100);
-   });
+//    // queue 100ms wait
+//    origFn.call(browser.driver.controlFlow(), function() {
+//      return protractor.promise.delayed(100);
+//    });
  
-   return origFn.apply(browser.driver.controlFlow(), args);
- };
+//    return origFn.apply(browser.driver.controlFlow(), args);
+//  };
 
 describe('Testando Módulo Produto:',()=>{
     let prodPage : ProdutoPage;
