+++ src/app/components/funcionario/cadastro-funcionario/cadastro-funcionario.component.html
@@ -116,7 +116,7 @@
 <mat-divider></mat-divider>
 <div style="margin-top: 10px;">
     <mat-action-row class="offset-md-10" style="padding-top: 5px;">
-      <button mat-button color="warn" routerLink="/home">Cancelar</button>
+      <button mat-button color="warn" routerLink="/home/funcionarios">Cancelar</button>
       <button mat-button color="primary" [disabled]="!f.valid" (click)="Salvar(f)">Concluir</button>
     </mat-action-row>
     </div>
