+++ src/app/components/funcionario/lista-funcionario/lista-funcionario.component.html
@@ -1,46 +1,64 @@
-<mat-card class="container">
-    <mat-card-content class="barra">
-  
-  <div class="container-table">
-      <div class="mat-elevation-z8">
-          <table class="table table-bordered table-hover">
-            <thead>
-              <tr>
-                <th id="id">#</th>
-                <th>Nome</th>
-                <th>Status</th>
-              </tr>
-            </thead>
-            <tbody>
-              <tr *ngFor="let funcionario of funcionarios">
-                <td>{{ funcionario.id }}</td>
-                <td>{{ funcionario.nome }}</td>
-                <td>{{ formatarStatus(funcionario.enabled) }}</td>
-                <td class="acoes"><button mat-raised-button color="primary" (click)="editar(funcionario.id)">Detalhar</button>
-                {{' '}}<button mat-raised-button *ngIf="funcionario.enabled" color="warn" (click)="deletar(funcionario.id)">Inativar</button>
-                {{' '}}<button mat-raised-button *ngIf="!funcionario.enabled" color="primary" (click)="reativar(funcionario.id)">Reativar</button></td>
-              </tr>
-             
-            </tbody>
-          </table>
-        
-          <div class="row justify-content-end">
-            <div class="col-6">
-              <mat-paginator [pageSizeOptions]="[5, 10, 25]"></mat-paginator>
-            </div>
-            <div class="fab-bottom-right">
-              <button id="btn-add" mat-fab routerLink="../cadastrofuncionario" color="primary"><mat-icon>add</mat-icon></button>
-            </div>
-          </div>
-        </div> 
-      </div> 
-  
-    </mat-card-content>
-  </mat-card>
-    <ng-template #deleteModal>
-      <div class="modal-body text-center">
-        <p>Tem certeza que deseja Inativar a funcionario?</p>
-        <button type="button" class="btn btn-default" (click)="confirm()">Sim</button>
-        <button type="button" class="btn btn-primary" (click)="decline()">Não</button>
+<div class="container-table">
+  <div class="mat-elevation-z8">    
+    <div class="row">
+      <div class="col-8">
+        <div class="titulo-table">
+         <h6 class="titulo">Funcionários</h6>
+        </div>
       </div>
-    </ng-template>
+      <div class="col-4">
+        <div class="filtro-table">
+          <mat-form-field class="filtro">
+            <mat-label>Busca: </mat-label>
+            <input matInput name="search" [(ngModel)]="filter">
+          </mat-form-field>
+        </div>
+      </div>
+    </div> 
+      <table class="table table-bordered table-hover">
+        <thead>
+          <tr>
+            
+            <th id="id">#</th>
+            <th (click)="sort('nome')">Nome
+              <span *ngIf="key =='nome'">{{reverse ? '▲' : '▼'}}</span>
+            </th>
+            <th>Status</th>
+            <th>Ações</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr *ngFor="let funcionario of funcionarios |filter:filter| orderBy: key : reverse | paginate: { itemsPerPage: 5, currentPage: p }">
+            <td>{{ funcionario.id }}</td>
+            <td>{{ funcionario.nome }}</td>
+            <td>{{ formatarStatus(funcionario.enabled) }}</td>
+            <td class="acoes"><button mat-raised-button color="primary"
+                (click)="editar(funcionario.id)">Detalhar</button>
+              {{' '}}<button mat-raised-button *ngIf="funcionario.enabled" color="warn"
+                (click)="deletar(funcionario.id)">Inativar</button>
+              {{' '}}<button mat-raised-button *ngIf="!funcionario.enabled" color="primary"
+                (click)="reativar(funcionario.id)">Reativar</button></td>
+          </tr>
+        </tbody>
+      </table>
+      <div class="row justify-content-center">
+        <pagination-controls (pageChange)="p = $event" previousLabel="Anterior"
+        nextLabel="Próximo"></pagination-controls>
+      </div>
+      <div class="row justify-content-end">
+        <div class="fab-bottom-right">
+          <button id="btn-add" mat-fab routerLink="../cadastrofuncionario" color="primary">
+            <mat-icon>add</mat-icon>
+          </button>
+        </div>
+      </div>
+    </div> 
+</div>
+
+<ng-template #deleteModal>
+  <div class="modal-body text-center">
+    <p>Tem certeza que deseja Inativar a funcionario?</p>
+    <button type="button" class="btn btn-default" (click)="confirm()">Sim</button>
+    <button type="button" class="btn btn-primary" (click)="decline()">Não</button>
+  </div>
+</ng-template>
