+++ src/app/components/marca/lista-marca/lista-marca.component.html
@@ -1,19 +1,34 @@
-<mat-card class="container">
-    <mat-card-content class="barra">
-  
-  <div class="container-table">
+<div class="container-table">
       <div class="mat-elevation-z8">
+        <div class="row">
+          <div class="col-8">
+            <div class="titulo-table">
+             <h6 class="titulo">Marcas</h6>
+            </div>
+          </div>
+          <div class="col-4">
+            <div class="filtro-table">
+              <mat-form-field class="filtro">
+                <mat-label>Busca: </mat-label>
+                <input matInput name="search" [(ngModel)]="filter">
+              </mat-form-field>
+            </div>
+          </div>
+        </div>
           <table class="table table-bordered table-hover">
             <thead>
               <tr>
                 <th id="id">#</th>
-                <th>Nome</th>
+                <th (click)="sort('nome')">Nome
+                  <span *ngIf="key =='nome'">{{reverse ? '▲' : '▼'}}</span>
+                </th>
                 <th>descrição</th>
+                <th>Ações</th>
   
               </tr>
             </thead>
             <tbody>
-              <tr *ngFor="let marca of marcas">
+              <tr *ngFor="let marca of marcas |filter:filter| orderBy: key : reverse | paginate: { itemsPerPage: 5, currentPage: p }">
                 <td>{{marca.id}}</td>
                 <td>{{marca.nome}}</td>
                 <td>{{marca.descricao}}</td>
@@ -34,9 +49,6 @@
           </div>
         </div> 
       </div> 
-  
-    </mat-card-content>
-  </mat-card>
     <ng-template #deleteModal>
       <div class="modal-body text-center">
         <p>Tem certeza que deseja excuir a marca?</p>
